requirements.txt
Flask
python-dotenv
pyyaml
Flask-SQLAlchemy
Flask-reCaptcha
Flask-Limiter
Flask-Mail
